+++ scripts/INSTALL_python_packages.sh
@@ -67,3 +67,4 @@ done
 
 $PYTHON -m pip install $PIPOPTIONS --upgrade pip wheel setuptools
 $PYTHON -m pip install $PIPOPTIONS --only-binary=numpy,scipy,matplotlib numpy scipy matplotlib nose coverage docopt
+$PYTHON -m pip install $PIPOPTIONS jupyter spyder==3.2