+++ scripts/UPDATE.sh
@@ -132,17 +132,16 @@ SuperBuild(){
   
   cd buildVM
   cmake ../SIRF-SuperBuild \
-      -DCMAKE_INSTALL_PREFIX=${SIRF_INSTALL_PATH} \
-	  -USIRF_URL -USIRF_TAG -USTIR_URL -USTIR_TAG \
-	  -UGadgetron_URL -UGadgetron_TAG -UISMRMRD_URL \
-	  -UISMRMRD_TAG \
-	  -DUSE_SYSTEM_SWIG=On \
-	  -DUSE_SYSTEM_Boost=On \
-	  -DUSE_SYSTEM_Armadillo=On \
-	  -DUSE_SYSTEM_FFTW3=On \
-	  -DUSE_SYSTEM_HDF5=ON \
-	  -DBUILD_siemens_to_ismrmrd=On \
-	  -DDEVEL_BUILD=OFF
+        -DCMAKE_INSTALL_PREFIX=${SIRF_INSTALL_PATH} \
+        -U\*_URL -U\*_TAG \
+        -DUSE_SYSTEM_SWIG=On \
+        -DUSE_SYSTEM_Boost=On \
+        -DUSE_SYSTEM_Armadillo=On \
+        -DUSE_SYSTEM_FFTW3=On \
+        -DUSE_SYSTEM_HDF5=ON \
+        -DBUILD_siemens_to_ismrmrd=On \
+        -DUSE_ITK=ON \
+        -DDEVEL_BUILD=OFF
   make -j${num_parallel}
 
   if [ ! -f ${SIRF_INSTALL_PATH}/share/gadgetron/config/gadgetron.xml ]